Run library.lua as part of ll::Interpreter construction
