+++ core/include/lluvia/core/impl/.gitignore
@@ -1,0 +1,1 @@
+LuaLibrary.h
