+++ core/include/lluvia/core/impl/LuaLibrary.h.in
@@ -1,0 +1,23 @@
+/**
+@file       InterpreterLibrary.h
+@brief      Lua library loaded into the ll::Interpreter
+@copyright  2018, Juan David Adarve Bermudez. See AUTHORS for more details.
+            Distributed under the Apache-2 license, see LICENSE for more details.
+*/
+
+#ifndef LLUVIA_CORE_IMPL_LUA_LIBRARY_H_
+#define LLUVIA_CORE_IMPL_LUA_LIBRARY_H_
+
+#include <string>
+
+namespace ll {
+namespace impl {
+
+constexpr const auto LUA_LIBRARY_SRC = R"(
+${LL_LUA_LIBRARY_SRC}
+)";
+
+} // namespace impl
+} // namespace ll
+
+#endif // LLUVIA_CORE_IMPL_LUA_LIBRARY_H_
